attribution_analysis/requirements.txt
# attribution_analysis/requirements.txt
baostock
tushare
futu-api
requests
python-dotenv
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
pyecharts>=2.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
